Add container load checkpoint error reporting + refactor
