+++ deepspeed/module_inject/containers/distil_bert.py
@@ -71,6 +71,3 @@ class HFDistilBertLayerPolicy(TransformerPolicy):
                attention_layernorm.bias, \
                transformer_layernorm.weight, \
                transformer_layernorm.bias
-
-    def get_param_names(self):
-        pass
