+++ deepspeed/module_inject/containers/gpt2.py
@@ -50,6 +50,3 @@ class HFGPT2LayerPolicy(TransformerPolicy):
                self.client_module.ln_2.bias, \
                self.client_module.ln_1.weight, \
                self.client_module.ln_1.bias
-
-    def get_param_names(self):
-        pass
